Resolve: linter issues from local machine
